+++ src/frontend/config.test.ts
@@ -1,25 +1,10 @@
 import { makeCtxConfig } from "../handler-context/context-builder";
-import {
-  getContractAndVendorName,
-  getContractPeriods,
-  getContracts,
-  getDashboardExtract,
-} from "./config";
-import { AthenaQueryExecutor } from "../shared/utils/athenaV3";
+import { getContractAndVendorName, getContracts } from "./config";
 import { describe } from "node:test";
 import { fetchS3 } from "../shared/utils";
 
 jest.mock("../handler-context/context-builder");
 const mockedMakeCtxConfig = makeCtxConfig as jest.Mock;
-
-jest.mock("../shared/utils/athenaV3");
-const MockedAthenaQueryExecutor = AthenaQueryExecutor as jest.MockedClass<
-  typeof AthenaQueryExecutor
->;
-const mockedAthenaQueryExecutorFetchResults = jest.fn();
-MockedAthenaQueryExecutor.mockReturnValue({
-  fetchResults: mockedAthenaQueryExecutorFetchResults,
-} as any);
 
 jest.mock("../shared/utils");
 const mockedFetchS3 = fetchS3 as jest.Mock;
@@ -27,13 +12,11 @@ const mockedFetchS3 = fetchS3 as jest.Mock;
 describe("frontend config", () => {
   let givenContractsConfig;
   let givenServicesConfig;
-  let givenQueryResults;
   let contractId: string;
 
   beforeEach(() => {
     process.env = {
-      QUERY_RESULTS_BUCKET: "given query results bucket",
-      DATABASE_NAME: "given database name",
+      STORAGE_BUCKET: "given storage bucket",
     };
 
     contractId = "1";
@@ -56,16 +39,9 @@ describe("frontend config", () => {
       contracts: givenContractsConfig,
     });
 
-    givenQueryResults = {
-      ColumnInfos: [],
-      ResultRows: [],
-      ResultSetMetadata: { ColumnInfo: [] },
-      Rows: [
-        { Data: [{ VarCharValue: "month" }, { VarCharValue: "year" }] },
-        { Data: [{ VarCharValue: "03" }, { VarCharValue: "2023" }] },
-      ],
-    };
-    mockedAthenaQueryExecutorFetchResults.mockResolvedValue(givenQueryResults);
+    mockedFetchS3.mockResolvedValue(
+      '{"month":"03","year":"2023","contract_id":"1"}'
+    );
   });
 
   describe("getContracts", () => {
@@ -95,70 +71,6 @@ describe("frontend config", () => {
       await expect(getContractAndVendorName("2")).rejects.toThrowError(
         "No contract found"
       );
-    });
-  });
-
-  describe("getContractPeriods", () => {
-    test("should return the month, year and prettyMonth", async () => {
-      // Act
-      const result = await getContractPeriods(contractId);
-      // Assert
-      expect(result).toEqual([
-        { month: "03", prettyMonth: "Mar", year: "2023" },
-      ]);
-    });
-
-    test("should throw an error if there is no QUERY_RESULTS_BUCKET", async () => {
-      // Act
-      delete process.env.QUERY_RESULTS_BUCKET;
-      // Assert
-      await expect(getContractPeriods(contractId)).rejects.toThrowError(
-        "No QUERY_RESULTS_BUCKET defined in this environment"
-      );
-    });
-
-    test("should throw an error if there is no DATABASE_NAME", async () => {
-      // Act
-      delete process.env.DATABASE_NAME;
-      // Assert
-      await expect(getContractPeriods(contractId)).rejects.toThrowError(
-        "No DATABASE_NAME defined in this environment"
-      );
-    });
-
-    test("should throw an error if there are no results in result set", async () => {
-      // Act
-      mockedAthenaQueryExecutorFetchResults.mockResolvedValue({});
-      // Assert
-      await expect(getContractPeriods(contractId)).rejects.toThrowError(
-        "No results in result set"
-      );
-    });
-
-    test("should throw an error if there is data missing", async () => {
-      // Act
-      mockedAthenaQueryExecutorFetchResults.mockResolvedValue({
-        ColumnInfos: [],
-        ResultRows: [],
-        ResultSetMetadata: { ColumnInfo: [] },
-        Rows: [
-          { Data: [{ VarCharValue: "month" }, { VarCharValue: "year" }] },
-          { Data: [{ VarCharValue: "03" }, {}] }, // Missing year
-        ],
-      });
-      // Assert
-      await expect(getContractPeriods(contractId)).rejects.toThrowError(
-        "Some data was missing"
-      );
-    });
-  });
-
-  describe("getDashboardExtract", () => {
-    test("Should parse multi-line JSON into object", async () => {
-      const mockedS3FileText = '{"a":"b","c":"d"}\n{"e":"f"}\n{"g":"h"}';
-      const expectedExtract = [{ a: "b", c: "d" }, { e: "f" }, { g: "h" }];
-      mockedFetchS3.mockReturnValue(mockedS3FileText);
-      expect(await getDashboardExtract()).toEqual(expectedExtract);
     });
   });
 });
