+++ src/frontend/handlers/invoice.test.ts
@@ -2,13 +2,18 @@ import supertest from "supertest";
 import { app } from "../app";
 import { makeCtxConfig } from "../../handler-context/context-builder";
 import { initApp } from "../init-app";
+import { fetchS3 } from "../../shared/utils";
 
 jest.mock("../../handler-context/context-builder");
 const mockedMakeCtxConfig = makeCtxConfig as jest.Mock;
 
+jest.mock("../../shared/utils");
+const mockedFetchS3 = fetchS3 as jest.Mock;
+
 describe("invoice handler", () => {
   let givenContractsConfig;
   let givenServicesConfig;
+  let givenExtractResults;
   let contractId: string;
 
   beforeEach(() => {
@@ -28,11 +33,16 @@ describe("invoice handler", () => {
         contract_id: contractId,
       },
     ];
-    // Arrange
     mockedMakeCtxConfig.mockResolvedValue({
       services: givenServicesConfig,
       contracts: givenContractsConfig,
     });
+
+    givenExtractResults =
+      '{"month":"10","year":"2023","contract_id":"1"}\n' +
+      '{"month":"03","year":"2023","contract_id":"1"}\n' +
+      '{"month":"04","year":"2023","contract_id":"1"}';
+    mockedFetchS3.mockResolvedValue(givenExtractResults);
   });
 
   test("Page displays invoice info", async () => {
@@ -45,5 +55,17 @@ describe("invoice handler", () => {
     expect(response.text).toContain("Contracts");
     expect(response.text).toContain("C01234 - Vendor One");
     expect(response.text).toMatchSnapshot();
+  });
+
+  test("Page displays 'no data' message", async () => {
+    givenExtractResults = "";
+    mockedFetchS3.mockResolvedValue(givenExtractResults);
+
+    const request = supertest(app);
+    const response = await request.get(
+      "/invoice?contract_id=1&year=2023&month=03"
+    );
+    expect(response.status).toBe(200);
+    expect(response.text).toContain("Invoice and events missing");
   });
 });
