+++ src/frontend/handlers/invoices.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from "express";
-import { getContractAndVendorName, getContractPeriods } from "../config";
+import { getContractAndVendorName } from "../config";
+import { getContractPeriods } from "../extract-helper";
 
 export const getInvoicesHandler: RequestHandler<
   unknown,
@@ -7,7 +8,7 @@ export const getInvoicesHandler: RequestHandler<
   unknown,
   { contract_id: string }
 > = async (request, response) => {
-  const [config, periods] = await Promise.all([
+  const [{ contractName, vendorName }, periods] = await Promise.all([
     getContractAndVendorName(request.query.contract_id),
     getContractPeriods(request.query.contract_id),
   ]);
@@ -15,8 +16,8 @@ export const getInvoicesHandler: RequestHandler<
   response.render("invoices.njk", {
     contract: {
       id: request.query.contract_id,
-      name: config.contractName,
-      vendorName: config.vendorName,
+      name: contractName,
+      vendorName,
     },
     periods,
   });
